Consolidate types and general cleanup for schema-tools
