+++ packages/@tinacms/schema-tools/src/types/index.ts
@@ -1,7 +1,558 @@
 /**
 
-
-
 */
+import type { FC } from 'react'
 
-export * from './SchemaTypes'
+import { TinaSchema } from '../schema'
+import { Config } from '../types'
+
+/**
+ * @deprecated use Config instead
+ */
+export type TinaCloudSchemaConfig<DeleteMe = undefined> = Config
+
+export type { Config }
+
+export type UIField<F extends UIField = any, Shape = any> = {
+  // name?: string
+  label?: string
+  description?: string
+  // TODO type component
+  component?: FC<any> | string | null
+  // inlineComponent?: FC<any>
+  parse?: (value: Shape, name: string, field: F) => any
+  format?: (value: Shape, name: string, field: F) => any
+  validate?(
+    value: Shape,
+    allValues: any,
+    meta: any,
+    field: UIField<F, Shape>
+  ): string | object | undefined | void
+  /**
+   * @deprecated use `defaultItem` at the collection level instead
+   */
+  defaultValue?: Shape
+}
+
+export interface TinaCloudSchema<WithNamespace extends boolean, Store = any> {
+  collections: TinaCloudCollection<WithNamespace>[]
+  /**
+   * @deprecated use `defineConfig` in a config.{js,ts} file instead
+   */
+  config?: Config
+}
+export type TinaCloudSchemaBase = TinaCloudSchema<false>
+export type TinaCloudSchemaEnriched = TinaCloudSchema<true>
+
+/**
+ * As part of the build process, each node is given a `path: string[]` key
+ * to help with namespacing type names, this is added as part of the
+ * createTinaSchema step
+ */
+export interface TinaCloudSchemaWithNamespace {
+  collections: TinaCloudCollection<true>[]
+  config?: Config
+  namespace: string[]
+}
+
+export type TinaCloudCollection<WithNamespace extends boolean> =
+  | CollectionFields<WithNamespace>
+  | CollectionTemplates<WithNamespace>
+
+export type TinaCloudCollectionBase = TinaCloudCollection<false>
+export type TinaCloudCollectionEnriched = TinaCloudCollection<true>
+
+type FormatType = 'json' | 'md' | 'markdown' | 'mdx'
+
+type Document = {
+  _sys: {
+    title?: string
+    template: string
+    breadcrumbs: string[]
+    path: string
+    basename: string
+    relativePath: string
+    filename: string
+    extension: string
+  }
+}
+export type TinaIndex = {
+  name: string
+  fields: {
+    name: string
+  }[]
+}
+
+export interface UICollection {
+  /**
+   * Customize the way filenames are generated during content creation
+   */
+  filename?: {
+    /**
+     * A callback which receives form values as an argument. The return value
+     * here will be used as the filename (the extension is not necessary)
+     *
+     * eg:
+     * ```ts
+     * slugify: (values) => values.title.toLowerCase().split(" ").join("-")
+     * ```
+     */
+    slugify?: (values: Record<string, any>) => string
+    /**
+     * When set to `true`, editors won't be able to modify the filename
+     */
+    readonly?: boolean
+  }
+
+  allowedActions?: {
+    create?: boolean
+    delete?: boolean
+  }
+
+  /**
+   * Forms for this collection will be editable from the global sidebar rather than the form panel
+   */
+  global?: boolean | { icon?: any; layout: 'fullscreen' | 'popup' }
+  /**
+   * Provide the path that your document is viewable on your site
+   *
+   * eg:
+   * ```ts
+   * router: ({ document }) => {
+   *   return `blog-posts/${document._sys.filename}`;
+   * }
+   * ```
+   */
+  router?: (args: {
+    document: Document
+    collection: TinaCloudCollection<true>
+  }) => string | undefined
+}
+
+export type DefaultItem<ReturnType> = ReturnType | (() => ReturnType)
+
+interface BaseCollection {
+  label?: string
+  name: string
+  path: string
+  defaultItem?: DefaultItem<Record<string, any>>
+  indexes?: TinaIndex[]
+  format?: FormatType
+  /**
+   * This format will be used to parse the markdown frontmatter
+   */
+  frontmatterFormat?: 'yaml' | 'toml' | 'json'
+  /**
+   * The delimiters used to parse the frontmatter.
+   */
+  frontmatterDelimiters?: [string, string] | string
+  ui?: UICollection
+  match?: string
+}
+
+export type CollectionTemplates<WithNamespace extends boolean> =
+  WithNamespace extends true
+    ? CollectionTemplatesWithNamespace<WithNamespace>
+    : CollectionTemplatesInner<WithNamespace>
+
+export type TinaTemplate = Template<false>
+
+interface CollectionTemplatesInner<WithNamespace extends boolean>
+  extends BaseCollection {
+  templates: Template<WithNamespace>[]
+  fields?: undefined
+}
+export interface CollectionTemplatesWithNamespace<WithNamespace extends boolean>
+  extends BaseCollection {
+  templates: Template<WithNamespace>[]
+  fields?: undefined
+  references?: ReferenceType<WithNamespace>[]
+  namespace: WithNamespace extends true ? string[] : undefined
+}
+
+type CollectionFields<WithNamespace extends boolean> =
+  WithNamespace extends true
+    ? CollectionFieldsWithNamespace<WithNamespace>
+    : CollectionFieldsInner<WithNamespace>
+
+export interface CollectionFieldsWithNamespace<WithNamespace extends boolean>
+  extends BaseCollection {
+  fields: TinaFieldInner<WithNamespace>[]
+  templates?: undefined
+  references?: ReferenceType<WithNamespace>[]
+  namespace: string[]
+}
+
+interface CollectionFieldsInner<WithNamespace extends boolean>
+  extends BaseCollection {
+  fields: TinaFieldInner<WithNamespace>[]
+  templates?: undefined
+}
+
+export type TinaFieldInner<WithNamespace extends boolean> =
+  | ScalarType<WithNamespace>
+  | ObjectType<WithNamespace>
+  | ReferenceType<WithNamespace>
+  | RichType<WithNamespace>
+
+export type TinaFieldBase = TinaFieldInner<false>
+export type TinaFieldEnriched = TinaFieldInner<true> & {
+  /**
+   * The parentTypename will always be an object type, either the type of a
+   * collection (ie. `Post`) or of an object field (ie. `PageBlocks`).
+   */
+  parentTypename?: string
+}
+
+export interface TinaField {
+  name: string
+  label?: string
+  description?: string
+  required?: boolean
+  indexed?: boolean
+  // list?: boolean
+  /**
+   * Any items passed to the UI field will be passed to the underlying field.
+   * NOTE: only serializable values are supported, so functions like `validate`
+   * will be ignored.
+   */
+  ui?: Record<string, any>
+}
+
+type ScalarType<WithNamespace extends boolean> = WithNamespace extends true
+  ? ScalarTypeWithNamespace
+  : ScalarTypeInner
+
+export type Option =
+  | string
+  | {
+      label?: string
+      icon?: FC
+      value: string
+    }
+type ScalarTypeInner = TinaField &
+  TinaScalarField & {
+    options?: Option[]
+  }
+type ScalarTypeWithNamespace = TinaField &
+  TinaScalarField & {
+    options?: Option[]
+    namespace: string[]
+  }
+type TinaScalarField =
+  | StringField
+  | BooleanField
+  | DateTimeField
+  | NumberField
+  | ImageField
+
+type StringField =
+  | {
+      type: 'string'
+      isBody?: boolean
+      list?: false
+      isTitle?: boolean
+      ui?: UIField<any, string>
+    }
+  | {
+      type: 'string'
+      isBody?: boolean
+      list: true
+      isTitle?: never
+      ui?: UIField<any, string[]> & { defaultItem?: DefaultItem<string> }
+    }
+
+type BooleanField =
+  | {
+      type: 'boolean'
+      list?: false
+      ui?: object | UIField<any, boolean>
+    }
+  | {
+      type: 'boolean'
+      list: true
+      ui?: object | UIField<any, boolean[]>
+    }
+
+type NumberField =
+  | {
+      type: 'number'
+      list?: false
+      ui?: object | UIField<any, number>
+    }
+  | {
+      type: 'number'
+      list: true
+      ui?: object | UIField<any, number[]>
+    }
+
+type DateTimeField =
+  | {
+      type: 'datetime'
+      dateFormat?: string
+      timeFormat?: string
+      list?: false
+      ui?: object | UIField<any, string>
+    }
+  | {
+      type: 'datetime'
+      dateFormat?: string
+      timeFormat?: string
+      list: true
+      ui?: object | UIField<any, string[]>
+    }
+
+type ImageField =
+  | {
+      type: 'image'
+      list?: false
+      ui?: object | UIField<any, string>
+    }
+  | {
+      type: 'image'
+      list: true
+      ui?: object | UIField<any, string[]>
+    }
+
+export type ReferenceType<WithNamespace extends boolean> =
+  WithNamespace extends true ? ReferenceTypeWithNamespace : ReferenceTypeInner
+
+export type RichType<WithNamespace extends boolean> = WithNamespace extends true
+  ? RichTypeWithNamespace
+  : RichTypeInner
+export interface ReferenceTypeInner extends TinaField {
+  type: 'reference'
+  list?: boolean
+  reverseLookup?: { label: string; name: string }
+  collections: string[]
+  ui?: UIField<any, string[]>
+}
+export interface ReferenceTypeWithNamespace extends TinaField {
+  type: 'reference'
+  list?: boolean
+  collections: string[]
+  reverseLookup?: { label: string; name: string }
+  namespace: string[]
+  ui?: UIField<any, string[]>
+}
+
+export type RichTextTemplate<WithNamespace extends boolean> =
+  Template<WithNamespace> & {
+    inline?: boolean
+    match?: {
+      start: string
+      end: string
+      name?: string
+    }
+  }
+
+export interface RichTypeWithNamespace extends TinaField {
+  type: 'rich-text'
+  namespace: string[]
+  isBody?: boolean
+  list?: boolean
+  templates?: (Template<true> & {
+    inline?: boolean
+    match?: {
+      start: string
+      end: string
+      name?: string
+    }
+  })[]
+}
+
+export interface RichTypeInner extends TinaField {
+  type: 'rich-text'
+  isBody?: boolean
+  list?: boolean
+  parser?:
+    | {
+        type: 'markdown'
+        skipEscaping?: 'all' | 'html' | 'none'
+      }
+    | { type: 'mdx' }
+  templates?: (Template<false> & {
+    inline?: boolean
+    match?: {
+      start: string
+      end: string
+      name?: string
+    }
+  })[]
+}
+
+export type ObjectType<WithNamespace extends boolean> =
+  | ObjectTemplates<WithNamespace>
+  | ObjectFields<WithNamespace>
+
+type ObjectTemplates<WithNamespace extends boolean> = WithNamespace extends true
+  ? ObjectTemplatesWithNamespace<WithNamespace>
+  : ObjectTemplatesInner<WithNamespace>
+
+type ObjectTemplatesInner<WithNamespace extends boolean> =
+  | ObjectTemplatesInnerWithList<WithNamespace>
+  | ObjectTemplatesInnerWithoutList<WithNamespace>
+interface ObjectTemplatesInnerWithList<WithNamespace extends boolean>
+  extends ObjectTemplatesInnerBase<WithNamespace> {
+  list?: true
+  ui?:
+    | object
+    | ({
+        // TODO: we could Type item based on fields or templates? (might be hard)
+        itemProps?(item: Record<string, any>): {
+          key?: string
+          label?: string
+        }
+        defaultItem?: DefaultItem<Record<string, any>>
+      } & UIField<any, string>)
+}
+interface ObjectTemplatesInnerWithoutList<WithNamespace extends boolean>
+  extends ObjectTemplatesInnerBase<WithNamespace> {
+  list?: false
+  ui?: object | UIField<any, string>
+}
+
+interface ObjectTemplatesInnerBase<WithNamespace extends boolean>
+  extends TinaField {
+  type: 'object'
+  visualSelector?: boolean
+  required?: false
+  list?: boolean
+  templates: Template<WithNamespace>[]
+  fields?: undefined
+}
+
+interface ObjectTemplatesWithNamespace<WithNamespace extends boolean>
+  extends TinaField {
+  type: 'object'
+  visualSelector?: boolean
+  required?: false
+  ui?: UIField<any, Record<string, any>> & {
+    itemProps?(item: Record<string, any>): {
+      key?: string
+      label?: string
+    }
+    defaultItem?: DefaultItem<Record<string, any>>
+  }
+  list?: boolean
+  templates: Template<WithNamespace>[]
+  fields?: undefined
+  namespace: WithNamespace extends true ? string[] : undefined
+}
+
+type ObjectFields<WithNamespace extends boolean> = WithNamespace extends true
+  ? InnerObjectFieldsWithNamespace<WithNamespace>
+  : InnerObjectFields<WithNamespace>
+
+interface InnerObjectFields<WithNamespace extends boolean> extends TinaField {
+  type: 'object'
+  visualSelector?: boolean
+  required?: false
+  ui?: UIField<any, Record<string, any>> & {
+    itemProps?(item: Record<string, any>): {
+      key?: string
+      label?: string
+    }
+    defaultItem?: DefaultItem<Record<string, any>>
+  }
+  /**
+   * fields can either be an array of Tina fields, or a reference to the fields
+   * of a global template definition.
+   *
+   * You can only provide one of `fields` or `templates`, but not both.
+   */
+  fields: TinaFieldInner<WithNamespace>[]
+  templates?: undefined
+  list?: boolean
+}
+
+interface InnerObjectFieldsWithNamespace<WithNamespace extends boolean>
+  extends TinaField {
+  type: 'object'
+  visualSelector?: boolean
+  required?: false
+  ui?: UIField<any, Record<string, any>> & {
+    itemProps?(item: Record<string, any>): {
+      key?: string
+      label?: string
+    }
+    defaultItem?: DefaultItem<Record<string, any>>
+  }
+  /**
+   * fields can either be an array of Tina fields, or a reference to the fields
+   * of a global template definition.
+   *
+   * You can only provide one of `fields` or `templates`, but not both.
+   */
+  fields: TinaFieldInner<WithNamespace>[]
+  templates?: undefined
+  namespace: WithNamespace extends true ? string[] : undefined
+  list?: boolean
+}
+
+export type TinaCloudTemplateBase = Template<false>
+export type TinaCloudTemplateEnriched = Template<true>
+/**
+ * Templates allow you to define an object as polymorphic
+ */
+export type Template<WithNamespace extends boolean> = WithNamespace extends true
+  ? {
+      label: string
+      name: string
+      fields: TinaFieldInner<WithNamespace>[]
+      ui?: object | (UIField<any, any> & { previewSrc: string })
+      namespace: WithNamespace extends true ? string[] : undefined
+    }
+  : {
+      label: string
+      name: string
+      ui?: object | (UIField<any, any> & { previewSrc: string })
+      fields: TinaFieldInner<WithNamespace>[]
+    }
+
+// Builder types
+export type CollectionTemplateableUnion = {
+  namespace: string[]
+  type: 'union'
+  templates: Templateable[]
+}
+export type CollectionTemplateableObject = {
+  namespace: string[]
+  type: 'object'
+  visualSelector?: boolean
+  ui?: UIField<any, Record<string, any>> & {
+    itemProps?(item: Record<string, any>): {
+      key?: string
+      label?: string
+    }
+    defaultItem?: DefaultItem<Record<string, any>>
+  }
+  required?: false
+  template: Templateable
+}
+export type CollectionTemplateable =
+  | CollectionTemplateableUnion
+  | CollectionTemplateableObject
+
+export type Collectable = {
+  namespace: string[]
+  templates?: Templateable[]
+  fields?: TinaFieldEnriched[]
+  references?: ReferenceType<true>[]
+}
+
+export type Templateable = {
+  name: string
+  namespace: string[]
+  fields: TinaFieldEnriched[]
+  ui?: object
+}
+
+export type ResolveFormArgs = {
+  collection: TinaCloudCollection<true>
+  basename: string
+  template: Templateable
+  schema: TinaSchema
+}
+
+// export * from './config'
